+++ src/main/java/ua/lviv/lgs/controller/FacultyController.java
@@ -1,8 +1,11 @@
 package ua.lviv.lgs.controller;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,14 +34,22 @@ public class FacultyController {
     @GetMapping
     public String viewFacultyList(Model model) {
         List<Faculty> facultiesList = facultyService.findAll();
+        Map<Faculty, Integer> countApplicationsByFaculty = facultyService.countApplicationsByFaculty();
         model.addAttribute("faculties", facultiesList);
+        model.addAttribute("aplicationsByFaculty", countApplicationsByFaculty);
 
         return "facultyList";
     }
 
     @GetMapping("/create")
-    public String viewCreationForm(Model model) {
+    public String viewCreationForm(@RequestParam(name = "superRefererURI", required = false) String superRefererURI,
+                                   HttpServletRequest request, Model model) throws URISyntaxException {
         model.addAttribute("subjects", subjectService.findAll());
+        model.addAttribute("refererURI", new URI(request.getHeader("referer")).getPath());
+
+        if (superRefererURI != null) {
+            model.addAttribute("superRefererURI", superRefererURI);
+        }
 
         return "facultyCreator";
     }
@@ -49,6 +60,11 @@ public class FacultyController {
             Map<String, String> errors = ControllerUtils.getErrors(bindingResult);
             model.mergeAttributes(errors);
             model.addAttribute("subjects", subjectService.findAll());
+            model.addAttribute("refererURI", form.get("refererURI"));
+
+            if (form.get("superRefererURI") != "") {
+                model.addAttribute("superRefererURI", form.get("superRefererURI"));
+            }
 
             return "facultyCreator";
         }
@@ -58,15 +74,29 @@ public class FacultyController {
         if (facultyExists) {
             model.addAttribute("facultyExistsMessage", "Such a faculty already exists!");
             model.addAttribute("subjects", subjectService.findAll());
+            model.addAttribute("refererURI", form.get("refererURI"));
+
+            if (form.get("superRefererURI") != "") {
+                model.addAttribute("superRefererURI", form.get("superRefererURI"));
+            }
 
             return "facultyCreator";
         }
 
-        return "redirect:/faculty/";
+        if (form.get("superRefererURI") != "") {
+            return "redirect:" + form.get("superRefererURI");
+        }
+
+        return "redirect:" + form.get("refererURI");
     }
 
     @GetMapping("/edit")
     public String viewEditForm(@RequestParam("id") Faculty faculty, Model model) {
+        Map<Faculty, Integer> countApplicationsByFaculty = facultyService.countApplicationsByFaculty();
+        if (countApplicationsByFaculty.get(faculty) != 0) {
+            return "redirect:/403";
+        }
+
         model.addAttribute("faculty", faculty);
         model.addAttribute("subjects", subjectService.findAll());
 
@@ -74,7 +104,8 @@ public class FacultyController {
     }
 
     @PostMapping("/edit")
-    public String updateFaculty(@RequestParam("id") Faculty faculty, @RequestParam Map<String, String> form, @Valid Faculty updatedFaculty, BindingResult bindingResult, Model model) {
+    public String updateFaculty(@RequestParam("id") Faculty faculty, @RequestParam Map<String, String> form,
+                                @Valid Faculty updatedFaculty, BindingResult bindingResult, Model model) {
         if (bindingResult.hasErrors()) {
             Map<String, String> errors = ControllerUtils.getErrors(bindingResult);
             model.mergeAttributes(errors);
@@ -84,13 +115,25 @@ public class FacultyController {
             return "facultyEditor";
         }
 
-        facultyService.updateFaculty(updatedFaculty, form);
+        boolean facultyExists = !facultyService.updateFaculty(updatedFaculty, form);
+
+        if (facultyExists) {
+            model.addAttribute("facultyExistsMessage", "Such a faculty already exists!");
+            model.addAttribute("faculty", faculty);
+            model.addAttribute("subjects", subjectService.findAll());
+
+            return "facultyEditor";
+        }
 
         return "redirect:/faculty";
     }
 
     @GetMapping("/delete")
     public String deleteFaculty(@RequestParam("id") Faculty faculty) {
+        if (!faculty.getExamSubjects().isEmpty() && !faculty.getSpecialities().isEmpty()) {
+            return "redirect:/403";
+        }
+
         facultyService.deleteFaculty(faculty);
 
         return "redirect:/faculty";
